+++ src/main/java/com/ll/medium/domain/article/answer/entity/CommentForm.java
@@ -1,4 +1,4 @@
-package com.ll.medium.domain.article.answer;
+package com.ll.medium.domain.article.answer.entity;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
