+++ src/main/java/com/ll/medium/domain/article/recommend/entity/Recommend.java
@@ -1,4 +1,4 @@
-package com.ll.medium.domain.article.answer;
+package com.ll.medium.domain.article.recommend.entity;
 
 import com.ll.medium.domain.article.article.entity.Article;
 import com.ll.medium.domain.member.entity.Member;
@@ -6,25 +6,17 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Getter
 @Setter
-public class Comment {
+public class Recommend {
     @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    private String content;
-
-    private LocalDateTime createDate;
-
-    private LocalDateTime modifyDate;
 
     @ManyToOne
     private Article article;
 
     @ManyToOne
-    private Member author;
+    private Member member;
 }
